requirements.txt
urllib3==1.23
urljoin==1.0.0
requests==2.19.1
behave==1.2.6
PyHamcrest==1.9.0
requests-toolbelt==0.9.1
pycpfcnpj